+++ reducer/store/mino/mino.go
