+++ reducer/store/stage/stage.go
@@ -1,9 +1,9 @@
 package stage
 
 import (
-	l "github.com/tetris-CLI/line"
-	m "github.com/tetris-CLI/mino"
-	tm "github.com/tetris-CLI/tetrimino"
+	l "github.com/tetris-CLI/reducer/store/line"
+	m "github.com/tetris-CLI/reducer/store/mino"
+	tm "github.com/tetris-CLI/reducer/store/tetrimino"
 
 	config "github.com/tetris-CLI/config"
 )
@@ -25,11 +25,11 @@ func NewStage() Stage {
 //IsConflictedWith StageとTetriminoが競合しているかどうかの真偽値を返す
 func (stage *Stage) IsConflictedWith(tetrimino tm.Tetrimino) bool {
 	for _, mino := range tetrimino.Minos {
-		if mino.X < 0 || config.StageWidth - 1 < mino.X {
+		if mino.X < 0 || config.StageWidth-1 < mino.X {
 			return true
 		}
 
-		if mino.Y < 0 || config.StageHeight - 1 < mino.Y {
+		if mino.Y < 0 || config.StageHeight-1 < mino.Y {
 			return true
 		}
