+++ reducer/store/stage/stage_test.go
