+++ reducer/store/tetrimino/tetrimino_test.go
